moved navbar and topbar; added link to incore logo; removed unecessary div
